+++ lib/src/core/core.dart
@@ -1,8 +1,5 @@
 library cv;
 
-import 'dart:ffi' as ffi;
-
-import 'package:ffi/ffi.dart';
 import 'package:opencv_dart/src/core/rng.dart';
 
 import '../types/scalar.dart';
@@ -10,9 +7,6 @@ import '../types/types.dart';
 import '../types/point.dart';
 import '../types/mat_type.dart';
 import '../types/mat.dart';
-import '../types/rect.dart';
-import '../types/moments.dart';
-import '../types/extensions.dart';
 import '../constants.g.dart';
 import '../opencv.g.dart' as cvg;